Move edit mode related components to editMode directory
