+++ app/components/editMode/EditWidgetIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { FunctionComponent } from 'react';
 import type { IconProps } from 'semantic-ui-react';
-import { Icon } from './basic/index';
+import { Icon } from '../basic';
 
 export interface EditWidgetIconProps {
     onShowConfig: () => void;
