+++ app/components/shared/widgets/PageContent.tsx
@@ -3,13 +3,13 @@ import { isEmpty, map, noop, wrap } from 'lodash';
 import i18n from 'i18next';
 import WidgetsList from './WidgetsList';
 import { Confirm } from '../../basic';
-import type { AddWidgetModalProps } from '../../AddWidgetModal';
-import AddWidgetModal from '../../AddWidgetModal';
+import type { AddWidgetModalProps } from '../../editMode/AddWidgetModal';
+import AddWidgetModal from '../../editMode/AddWidgetModal';
 import './PageContent.css';
 import type { TabsProps } from '../../Tabs';
 import Tabs from '../../Tabs';
 import useWidgetsFilter from '../../useWidgetsFilter';
-import EditModeButton from '../../EditModeButton';
+import EditModeButton from '../../editMode/EditModeButton';
 import { useResettableState } from '../../../utils/hooks';
 import Consts from '../../../utils/consts';
 import EmptyContainerMessage from '../../EmptyContainerMessage';
